# api/src/main/java/fr/grozeille/db4all/api/dto/ColumnDefinitionDto.java
package fr.grozeille.db4all.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ColumnDefinitionDto {
    private String name;
    private String dataType;
}